requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.7
django-ace==1.0.11
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-dark==0.1.9
django-dob-widget==1.2.0
django-embed-video==1.4.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-snippet==1.0.1
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
filter==0.0.0.20200724
gunicorn==20.1.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.7
Pillow==8.2.0
playsound==1.2.2
plyer==2.0.0
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.4.3
pylint-plugin-utils==0.6
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.56.0
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1